+++ libs/ymir-core/include/ymir/state/state_vdp.hpp
@@ -214,6 +214,7 @@ struct VDPState {
             bool erase;
 
             uint64 cycleCount;
+            uint64 cyclesSpent;
         };
 
         struct NormBGLayerState {